pyproject.toml
[project]
name = "dora_arm_piper"
version = "0.1.0"
description = ""
authors = [
    {name = "Xiang Yang",email = "ryu-yang@qq.com"}
]
readme = "README.md"
requires-python = ">=3.8,<3.14"
dependencies = [
    "dora-rs-cli (>=0.3.11,<0.4.0)",
    "piper_sdk >= 0.0.8",
    "logging_mp",
]

[tool.setuptools]
packages = []

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"